Implemented arbitrary configuration/key files.
The old behavior is retained as the default behavior.

The `ConfigParser` is now kept in a global variable
`CONFIG_STORE` and used as a storage for configuration
values. It is accessed by function `get_config_store()`
and used in all places where configuration values are
used.

Function `set_config_file(filename)` changes the 
configuration file. When it is not called, the parser
uses "config.ini" as before.